requirements.txt
flask==2.0.3
pyodbc==5.2.0
python-dotenv==0.19.2
gunicorn==20.1.0
